refactor: HighlightTexts 기능 수정
